scripts/requirements.txt
requests
python-dotenv
odoorpc
webdavclient3